feat!: unify function and tool concepts
we have two entities regarding tools: the schema (originally
coming from mcp servers) that we return from list_tools, and the
schema that users can provide as part of a chat request. these are
separate things, and we also represent them separately in the wit
world. however, for most users, they will be the same. the only
difference is that the chat params one has a strict member, allowing
to specify strict schema adherence. however, we can choose not to
set this field when listing tools, and leave it up to the user to
set it.